refactor(zlibrary-app): rename internal _shared/download to book-download to avoid framework path collision
