+++ clis/zlibrary-app/_shared/snapshot/download-rules.js
@@ -28,7 +28,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'No download request was initiated — the CDP Fetch event never fired',
     hypothesis: 'Z-Library changed the download initiation flow (link selector, button click, or JS event binding)',
     hints: [
-      { file: '_shared/download/link.js', reason: 'Check link selector and download URL extraction' },
+      { file: '_shared/book-download/link.js', reason: 'Check link selector and download URL extraction' },
       { file: 'booklist-download.js', reason: 'Check download initiation logic and click sequence' },
     ],
   },
@@ -37,7 +37,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'Download request returned HTTP 204 No Content — gate/block with no message body',
     hypothesis: 'Z-Library added or tightened a request gate (token expiry, session check, rate-limit pre-check)',
     hints: [
-      { file: '_shared/download/transport.js', reason: 'Check request-stage header injection and token/session gate handling' },
+      { file: '_shared/book-download/transport.js', reason: 'Check request-stage header injection and token/session gate handling' },
     ],
   },
   'DL_STATUS_403': {
@@ -55,7 +55,7 @@ var ROOT_CAUSE_MAP = {
     hypothesis: 'Z-Library introduced or tightened rate-limiting on download endpoints',
     hints: [
       { file: 'booklist-download.js', reason: 'Add rate-limit backoff or retry strategy' },
-      { file: '_shared/download/transport.js', reason: 'Check rate-limit header parsing (Retry-After)' },
+      { file: '_shared/book-download/transport.js', reason: 'Check rate-limit header parsing (Retry-After)' },
     ],
   },
   'NO_CDN_REDIRECT': {
@@ -63,7 +63,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'Request chain exists but final hop is not a CDN host — expected CDN redirect chain broken',
     hypothesis: 'Z-Library changed CDN provider or redirect logic (new intermediate hop or direct download)',
     hints: [
-      { file: '_shared/download/transport.js', reason: 'Check redirect tracking and CDN URL pattern matching' },
+      { file: '_shared/book-download/transport.js', reason: 'Check redirect tracking and CDN URL pattern matching' },
     ],
   },
   'CDN_HTML_BLOCK': {
@@ -71,7 +71,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'CDN returned HTML block page instead of file content — download limit or captcha wall',
     hypothesis: 'Z-Library added or modified block page (quota exceeded, suspicious activity, captcha challenge)',
     hints: [
-      { file: '_shared/download/workflow.js', reason: 'Check block-page validation and quota detection logic' },
+      { file: '_shared/book-download/workflow.js', reason: 'Check block-page validation and quota detection logic' },
     ],
   },
   'MD5_MISMATCH': {
@@ -79,8 +79,8 @@ var ROOT_CAUSE_MAP = {
     summary: 'Downloaded file MD5 does not match CDN-provided MD5 — file corruption or wrong content',
     hypothesis: 'CDN served stale or corrupted content; or MD5 extraction from CDN URL changed format',
     hints: [
-      { file: '_shared/download/contracts.js', reason: 'Check MD5 extraction and verification logic' },
-      { file: '_shared/download/workflow.js', reason: 'Check MD5 comparison and error handling' },
+      { file: '_shared/book-download/contracts.js', reason: 'Check MD5 extraction and verification logic' },
+      { file: '_shared/book-download/workflow.js', reason: 'Check MD5 comparison and error handling' },
     ],
   },
   'STREAM_ZERO_BYTES': {
@@ -88,7 +88,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'CDN returned HTTP 200 but stream delivered zero bytes — I/O or stream handling failure',
     hypothesis: 'CDP transport stream handling changed (response body not read, connection dropped without error)',
     hints: [
-      { file: '_shared/download/transport.js', reason: 'Check IO stream / Fetch stream handling and zero-byte detection' },
+      { file: '_shared/book-download/transport.js', reason: 'Check IO stream / Fetch stream handling and zero-byte detection' },
     ],
   },
   'CDN_NON_200': {
@@ -96,7 +96,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'CDN returned non-200 HTTP status — CDN-side error or redirect misconfiguration',
     hypothesis: 'CDN configuration changed (new error page, redirect loop, or upstream failure)',
     hints: [
-      { file: '_shared/download/transport.js', reason: 'Check CDN response handling and status code validation' },
+      { file: '_shared/book-download/transport.js', reason: 'Check CDN response handling and status code validation' },
     ],
   },
   'FIXTURE_INSUFFICIENT': {
@@ -104,7 +104,7 @@ var ROOT_CAUSE_MAP = {
     summary: 'Fixture schema-valid but lacks sufficient diagnostic data for root cause classification',
     hypothesis: 'Fixture recorder did not capture enough fields (response headers, error details, request chain)',
     hints: [
-      { file: '_shared/download/fixture.js', reason: 'Extend fixture recorder to capture missing diagnostic fields' },
+      { file: '_shared/book-download/fixture.js', reason: 'Extend fixture recorder to capture missing diagnostic fields' },
     ],
   },
 }
